Intent: streamline edge cache tests via consolidation Operators involved: [Self-organization] Affected invariants: [#8] Key changes: - Consolidate edge version cache scenarios into a single parametrized module - Provide fixture for graph and EdgeCacheManager reuse - Remove redundant per-scenario test modules Expected risks/dissonances: Reduced isolation of scenarios mitigated by parameterization and shared fixture Metrics: N/A Equivalence map: N/A
